# Conflicts:
#	src/main/java/lotto/LottoSimulation.java
#	src/main/java/lotto/domain/LottoMachine.java
#	src/main/java/lotto/domain/WinningNumber.java
#	src/test/java/lotto/domain/LottoMachineTest.java
#	src/test/java/lotto/domain/LottosTest.java

+++ src/main/java/lotto/domain/LottoRanking.java
@@ -33,7 +33,7 @@ public enum LottoRanking {
     }
 
     public static LottoRanking valueOf(Lotto lotto, WinningNumber winningNumber){
-        int matchedCount = winningNumber.getMatchedWinningNumberCount(lotto);
+        long matchedCount = winningNumber.getMatchedWinningNumberCount(lotto);
         boolean matchedBonus = winningNumber.matchBonusNumber(lotto);
 
         return Arrays.stream(LottoRanking.values())
